+++ src/BudgetBakers.Wallet.Net/Models/Account/CreateAccountRequest.cs
@@ -7,7 +7,7 @@ namespace BudgetBakers.Wallet.Net.Models.Account
         public required string Name { get; set; }
         public required CreatableAccountType AccountType { get; set; }
         public required string CurrencyCode { get; set; }
-        public required double InitialBalance { get; set; }
+        public required decimal InitialBalance { get; set; }
         public EntityColor? Color { get; set; }
         public string? BankAccountNumber { get; set; }
         public bool ExcludeFromStats { get; set; }
